Moved request to separate file
